requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
bcrypt>=4.0.1
psycopg2-binary>=2.9.9
pg8000==1.30.2